+++ pysgrs/breakers/hill.py
@@ -39,11 +39,13 @@ class StochasticHillClimbingBreaker(GenericLocalSearchBreaker):
             # Sample and score new state:
             next_state = self._score_state(self._next_state(**kwargs), text)
 
-            # Is next state, better?
+            # Is the next state, better?
             if next_state.score > self.current_state.score:
+                # Then go for it!
                 self._current_state = next_state
                 self._current_state.trial_counter = 0
             else:
+                # Look a bit further to see if there is something else:
                 self._current_state.trial_counter += 1
                 if self.current_state.trial_counter > max_trials:
                     break
@@ -57,12 +59,19 @@ class StochasticHillClimbingBreaker(GenericLocalSearchBreaker):
 
 def main():
     from pysgrs import scores
+    from pysgrs import toolbox
 
     t = """Bonjour tout le monde, on essaye des trucs, on teste, mais c'est pas absolu comme méthode...
-    Il sera sans doute nécessaire de s'appliquer d'avantage pour mieux comprendre les tenants et aboutissant
+    Il sera sans doute nécessaire de s'appliquer d'avantage pour mieux comprendre les tenants et aboutissants
     d'une telle méthodologie. En ajoutant du texte on parvient à améliorer le score et donc le critère de
-    ceovergence semble plus efficace, même si ça reste une illusion de pouvoir explore l'esapce des états de
-    manière exhaustive"""
+    convergence semble plus efficace, même si ça reste une illusion de pouvoir explorer l'esapce des états de
+    manière exhaustive. Mais bon, en rejouant plusieurs fois l'algorithme on parvient à trouver vingt-deux
+    des vingt-six caractères recherchés en moins de dix mille itérations et ça c'est déjà quelque chose
+    de positif et encourageant. Il est a noter que la taille, mais également le contenu du texte ont de
+    de l'importance. A cela s'ajoute également que la solution exacte n'est pas forcément celle qui possède
+    le plus haut score en terme de maximum de vraissemblance des digrammes."""
+
+    t = toolbox.AsciiCleaner.strip_accents(t)
 
     p = np.random.permutation(26)
     C = PermutationCipher(p)